removed session modification from main reducer - unused
