+++ client/src/SessionView/Map/MapUpload.js
@@ -1,29 +1,28 @@
-import { useContext, useCallback } from 'react'
+import { useCallback } from 'react'
 import { Map as MapIcon } from '@material-ui/icons'
+import { useQueryClient } from 'react-query'
 
 import ImageUpload from '../../shared/ImageUpload'
 import { useDomainErrors } from '../../shared/errorHandling'
-import { DispatchContext } from '../../state'
 
-function MapUpload({ sessionService, sessionId }) {
-  const dispatch = useContext(DispatchContext)
+function MapUpload({ sessionService, sessionId, onUpload }) {
   const { setError } = useDomainErrors()
+  const queryClient = useQueryClient()
 
   const upload = useCallback(
     async (file, previewUrl) => {
       try {
         const result = await sessionService.uploadMap(file, sessionId)
         if (result.ok) {
-          dispatch({
-            type: 'SetSessionMap',
-            payload: { sessionId, map: previewUrl },
-          })
+          onUpload(previewUrl)
+          // TODO move out
+          queryClient.invalidateQueries(['session', sessionId])
         }
       } catch (e) {
         setError(e)
       }
     },
-    [sessionId, dispatch, sessionService, setError],
+    [sessionId, onUpload, sessionService, setError, queryClient],
   )
 
   return (
