+++ client/src/SessionView/SessionView.js
@@ -14,7 +14,7 @@ import { useSessionContext } from './useSessionContext'
 import { Overview } from './Overview'
 import ShareButton from './ShareButton'
 import { EditButton } from './Edit'
-import Map from './Map'
+import { Map } from './Map'
 import SessionNavigation from './SessionNavigation'
 import DetailsForm from './DetailsForm'
 import LockForEdit from './LockForEdit'
